Rename renderer.ts to taskLogRenderer.ts
