+++ node-src/renderer/storybookInfo.ts
@@ -6,7 +6,7 @@ import {
 import { Context } from '../types';
 import { initial, pending, success } from '../ui/tasks/storybookInfo';
 import { runTask } from './engine';
-import { clackTaskLogRenderer } from './engine/clack/renderer';
+import { clackTaskLogRenderer } from './engine/clack/taskLogRenderer';
 import { getRenderer } from './engine/getRenderer';
 
 /**
